requirements.txt
requests
google-generativeai
flask
flask-cors
analyzer
dotenv